TS: StructureListFilters/ExperimentType
